Added custom rule to check descriptive links
